Fold status into get
