+++ cmd/workflow/workflow.go
@@ -16,7 +16,6 @@ import (
 	"github.com/smartcontractkit/cre-cli/cmd/workflow/simulate"
 	supported_chains "github.com/smartcontractkit/cre-cli/cmd/workflow/supported_chains"
 	"github.com/smartcontractkit/cre-cli/cmd/workflow/test"
-	"github.com/smartcontractkit/cre-cli/cmd/workflow/wfstatus"
 	"github.com/smartcontractkit/cre-cli/internal/runtime"
 )
 
@@ -29,7 +28,6 @@ func New(runtimeContext *runtime.Context) *cobra.Command {
 
 	workflowCmd.AddCommand(supported_chains.New(runtimeContext))
 	workflowCmd.AddCommand(activate.New(runtimeContext))
-	workflowCmd.AddCommand(wfstatus.New(runtimeContext))
 	workflowCmd.AddCommand(build.New(runtimeContext))
 	workflowCmd.AddCommand(convert.New(runtimeContext))
 	workflowCmd.AddCommand(delete.New(runtimeContext))
