+++ docs/cre_workflow_get.md
@@ -1,10 +1,10 @@
 ## cre workflow get
 
-Shows metadata for the workflow configured in workflow.yaml
+Show deployment health and recent execution for the workflow in workflow.yaml
 
 ### Synopsis
 
-Looks up the workflow whose name is configured for the selected --target in workflow.yaml and prints its metadata from the CRE platform. By default results are filtered to the workflow's configured deployment-registry; pass --all-registries to show matches from every registry.
+Looks up the workflow whose name is configured for the selected --target in workflow.yaml and prints deployment health and the most recent execution from the CRE platform. By default resolution is scoped to the workflow's configured deployment-registry; pass --all-registries to resolve across every registry.
 
 ```
 cre workflow get <workflow-folder-path> [optional flags]
@@ -15,13 +15,16 @@ cre workflow get <workflow-folder-path> [optional flags]
 ```
 cre workflow get ./my-workflow --target staging
   cre workflow get ./my-workflow --target staging --all-registries
+  cre workflow get ./my-workflow --target staging --output json
 ```
 
 ### Options
 
 ```
-      --all-registries   Do not filter results by the workflow's deployment-registry
+      --all-registries   Resolve the workflow across every registry instead of the configured deployment-registry
   -h, --help             help for get
+      --json             Output as JSON (shorthand for --output=json)
+      --output string    Output format: "json" prints JSON to stdout
 ```
 
 ### Options inherited from parent commands
@@ -40,4 +43,3 @@ cre workflow get ./my-workflow --target staging
 ### SEE ALSO
 
 * [cre workflow](cre_workflow.md)	 - Manages workflows
-
